Combine 3 fxml together and impletement converting page
