+++ AlgoDomino/build.fxbuild
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="ASCII"?>
 <anttasks:AntTask xmi:version="2.0" xmlns:xmi="http://www.omg.org/XMI" xmlns:anttasks="http://org.eclipse.fx.ide.jdt/1.0" buildDirectory="${project}/build">
   <deploy>
-    <application name="AlgoDonimo"/>
+    <application name="abc"/>
     <info/>
   </deploy>
   <signjar/>
