+++ AlgoDomino/src/application/ComparingController.java
@@ -1,0 +1,11 @@
+package application;
+
+import java.net.URL;
+import java.util.ResourceBundle;
+
+public class ComparingController {
+	
+	public void initialize(URL fxmlFileLocation, ResourceBundle resources) {
+
+  }
+}
