+++ AlgoDomino/src/application/MainController.java
@@ -1,0 +1,19 @@
+package application;
+
+import java.net.URL;
+import java.util.ResourceBundle;
+
+
+
+
+
+public class MainController{
+  
+
+
+    public void initialize(URL fxmlFileLocation, ResourceBundle resources) {
+
+
+    }
+    
+}
