--- AlgoDomino/src/unnc/cs/grape/view/controlStyle.css
@@ -1,7 +1,0 @@
-/*controlStyle.css */
- 
-.separator{
-    -fx-background-color: #FF0000;
-    -fx-background-radius: 2;
-}
-
